git-svn-id: https://active-ist.svn.sourceforge.net/svnroot/active-ist/libraries/jaus++/trunk@193 846840b6-6447-0410-a3b3-dec9255eb4eb

+++ 1.0/src/jaus/components/transport/jtcpclient.cpp
@@ -134,7 +134,7 @@ int JTCPClient::Send(const Jaus::Stream& msg)
     int result = 0;
     mTransportMutex.Enter();
     mSendStream.SetLength((unsigned int)(gNetworkHeader.size()));
-    mSendStream.SetWritePos(gNetworkHeader.size());
+    mSendStream.SetWritePos((unsigned int)gNetworkHeader.size());
     mSendStream.Write(msg);
     result = mTCP.Send(mSendStream);
     mTransportMutex.Leave();
